some extractions and tests
