+++ src/Paprika.Tests/DataPageTests.cs
@@ -7,7 +7,7 @@ using static Paprika.Tests.Values;
 
 namespace Paprika.Tests;
 
-public unsafe class DataPageTests
+public unsafe class DataPageTests : BasePageTests
 {
     private const byte RootLevel = 0;
 
@@ -19,7 +19,7 @@ public unsafe class DataPageTests
         var page = AllocPage();
         page.Clear();
 
-        var batch = new TestBatchContext(BatchId);
+        var batch = NewBatch(BatchId);
         var dataPage = new DataPage(page);
         var ctx = new SetContext(Key0, Balance0, Nonce0);
 
@@ -37,7 +37,7 @@ public unsafe class DataPageTests
         var page = AllocPage();
         page.Clear();
 
-        var batch = new TestBatchContext(BatchId);
+        var batch = NewBatch(BatchId);
 
         var dataPage = new DataPage(page);
         var ctx1 = new SetContext(Key0, Balance0, Nonce0);
@@ -57,7 +57,7 @@ public unsafe class DataPageTests
         var page = AllocPage();
         page.Clear();
 
-        var batch = new TestBatchContext(BatchId);
+        var batch = NewBatch(BatchId);
 
         var dataPage = new DataPage(page);
         var ctx1 = new SetContext(Key1a, Balance0, Nonce0);
@@ -81,7 +81,7 @@ public unsafe class DataPageTests
         var page = AllocPage();
         page.Clear();
 
-        var batch = new TestBatchContext(BatchId);
+        var batch = NewBatch(BatchId);
         var dataPage = new DataPage(page);
 
         const int count = 2 * 1024 * 1024;
@@ -103,78 +103,5 @@ public unsafe class DataPageTests
             dataPage.GetAccount(key, batch, out var account, RootLevel);
             account.Should().Be(new Account(i, i));
         }
-    }
-
-    [Test]
-    public void SetUp()
-    {
-        Pages.Clear();
-
-        foreach (var page in All)
-        {
-            Pages.Push(page);
-        }
-
-        All.Clear();
-    }
-
-
-    private static readonly Stack<Page> Pages = new();
-    private static readonly List<Page> All = new();
-
-    private static Page AllocPage()
-    {
-        if (Pages.TryPop(out var page))
-            return page;
-
-        const int slab = 16;
-
-        var memory = (byte*)NativeMemory.AlignedAlloc((UIntPtr)Page.PageSize * (nuint)slab, (UIntPtr)sizeof(long));
-
-        for (var i = 1; i < slab; i++)
-        {
-            var item = new Page(memory + Page.PageSize * i);
-            Pages.Push(item);
-            All.Add(item);
-        }
-
-        page = new Page(memory);
-        All.Add(page); // memo for reuse
-        return page;
-    }
-
-    internal class TestBatchContext : BatchContextBase
-    {
-        private readonly Dictionary<DbAddress, Page> _address2Page = new();
-
-        // data pages should start at non-null addresses
-        // 0-N is take by metadata pages
-        private uint _pageCount = 1U;
-
-        public TestBatchContext(uint batchId) : base(batchId, 0) { }
-
-        public override Page GetAt(DbAddress address) => _address2Page[address];
-
-        public override Page GetNewPage(out DbAddress addr, bool clear)
-        {
-            var page = AllocPage();
-            if (clear)
-                page.Clear();
-
-            page.Header.BatchId = BatchId;
-
-            addr = DbAddress.Page(_pageCount++);
-
-            _address2Page[addr] = page;
-
-            return page;
-        }
-
-        protected override void RegisterForFutureGC(Page page)
-        {
-            // NOOP
-        }
-
-        public override string ToString() => $"Batch context used {_pageCount} pages to write the data";
     }
 }
